Added comments, removed the obsolete IHaveViewModel.cs and SearchExtensions.cs files, and refactored tests, adding inner test classes and fakes into the Fakes folder.
