+++ AsyncMediator.Test/Fakes/ChainedEventHandler.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace AsyncMediator.Test
+{
+    public class ChainedEventHandler : IEventHandler<FakeEventTwoFromHandler>
+    {
+        public virtual Task Handle(FakeEventTwoFromHandler @event)
+        {
+            return Task.FromResult(6);
+        }
+    }
+}
