+++ AsyncMediator.Test/Fakes/FakeEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace AsyncMediator.Test
+{
+    public class FakeEvent : IDomainEvent
+    {
+        public int Id { get; set; }
+    }
+}
