+++ AsyncMediator.Test/Fakes/FakeEventFromHandler.cs
@@ -1,0 +1,7 @@
+namespace AsyncMediator.Test
+{
+    public class FakeEventFromHandler : IDomainEvent
+    {
+        public int Id { get; set; }
+    }
+}
