+++ AsyncMediator.Test/Fakes/FakeEventTwoFromHandler.cs
@@ -1,0 +1,7 @@
+namespace AsyncMediator.Test
+{
+    public class FakeEventTwoFromHandler : IDomainEvent
+    {
+        public int Id { get; set; }
+    }
+}
