+++ AsyncMediator.Test/Fakes/FakeRangeCriteria.cs
@@ -1,0 +1,8 @@
+namespace AsyncMediator.Test
+{
+    public class FakeRangeCriteria
+    {
+        public int MaxValue { get; set; }
+        public int MinValue { get; set; }
+    }
+}
