+++ AsyncMediator.Test/Fakes/FakeResult.cs
@@ -1,0 +1,9 @@
+namespace AsyncMediator.Test
+{
+    public class FakeResult
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string RelatedData { get; set; }
+    }
+}
