+++ AsyncMediator.Test/Fakes/HandlerWithoutEventsWithoutOrdering.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace AsyncMediator.Test
+{
+    public class HandlerWithoutEventsWithoutOrdering : IEventHandler<FakeEvent>
+    {
+        public virtual Task Handle(FakeEvent @event)
+        {
+            return Task.FromResult(5);
+        }
+    }
+}
