+++ AsyncMediator.Test/Fakes/SingleIdCriteria.cs
@@ -1,0 +1,7 @@
+namespace AsyncMediator.Test
+{
+    public class SingleIdCriteria
+    {
+        public int Id { get; set; }
+    }
+}
