+++ AsyncMediator.Test/Fakes/SingleNameCriteria.cs
@@ -1,0 +1,7 @@
+namespace AsyncMediator.Test
+{
+    public class SingleNameCriteria
+    {
+        public string Name { get; set; }
+    }
+}
