+++ AsyncMediator.Test/Fakes/TestCommand.cs
@@ -1,0 +1,7 @@
+namespace AsyncMediator.Test
+{
+    public class TestCommand : ICommand
+    {
+        public int Id { get; set; }
+    }
+}
