+++ Mediator/Interfaces/ICommand.cs
@@ -1,7 +1,7 @@
 namespace AsyncMediator
 {
     /// <summary>
-    /// A marker interface to indicate that the implementing object is a Command
+    /// A marker interface to indicate that the implementing object is a Command.
     /// </summary>
     public interface ICommand
     {
