+++ Mediator/Interfaces/IDomainEvent.cs
@@ -1,10 +1,9 @@
 namespace AsyncMediator
 {
     /// <summary>
-    /// A marker interface used to indicate that the object implementing it as an Event
+    /// A marker interface used to indicate that the object implementing it as an Event.
     /// </summary>
     public interface IDomainEvent
     {
-
     }
 }
