--- Mediator/Interfaces/IHaveViewModel.cs
@@ -1,7 +1,0 @@
-namespace AsyncMediator
-{
-    public interface IHaveViewModel<out TResult, in TSource>
-    {
-        TResult ToViewModel(TSource source);
-    }
-}
